'&*self.0' mystery
On line 21, I don't understand why a self-referencing struct requires an
extra '&'. On the hand, the 'Node' struct is not self-referencing so we can
write it as '*self.0'. When I remove the '&' from line 21, I get:

-----------------------------------------------------------------------------------------
error[E0507]: cannot move out of `self.0.0.1` which is behind a mutable reference
  --> src/main.rs:21:39
   |
21 |         while let Some((val, l, r)) = *self.0 {
   |                              -  -     ^^^^^^^ help: consider borrowing here: `&*self.0`
   |                              |  |
   |                              |  ...and here
   |                              data moved here
   |
   = note: move occurs because these variables have types that don't implement the `Copy` trait
-----------------------------------------------------------------------------------------

This is currently over my head.

Interestingly, when I change '*self.0' to '&*self.0' on line 31, the
code compiles fine.